+++ bot/handlers/authorization/customer_auth/confirm_password.py
@@ -68,13 +68,23 @@ async def confirm_password(message: Message, state: FSMContext):
             response.raise_for_status()
             response_data = response.json()
             access_token = response_data.get("access_token")
-            if not access_token:
+            refresh_token = response_data.get("refresh_token")
+            if not access_token and not refresh_token:
                 sent = await message.answer(str(InvalidTokenException(InvalidTokenException.__doc__)))
                 await asyncio.sleep(5)
                 await sent.delete()
                 return
             
             user_data = await Security.decode_jwt(access_token=access_token)
+
+            await client.post(
+                f"{settings.BASE_FASTAPI_URL}/tokens/",
+                json={
+                    "user_id": user_id,
+                    "access_token": access_token,
+                    "refresh_token": refresh_token
+                }
+            )  
 
         except httpx.HTTPStatusError:
             sent = await message.answer(
@@ -99,8 +109,8 @@ async def confirm_password(message: Message, state: FSMContext):
         if not is_change:
             await proceed_to_main_menu(user_data=user_data, message=message, state=state)
             await state.set_state(CustomerAuth.main_menu)
-            await state.update_data(user_data=user_data, access_token=access_token)
+            await state.update_data(user_data=user_data)
         
         else:
             await message.answer(Welcome.WELCOME, reply_markup=await BasicKeyboards.get_welcoming_kb())
-            await state.update_data(user_data=user_data, access_token=access_token)
+            await state.update_data(user_data=user_data)
